Refactor: Rename NetCoreFramework to NetCore
Refactor: Rename NetCoreFramework to NetCore
